# api.yaml
openapi: 3.0.1
info:
  title: Exam Portal API
  version: 1.0.0
servers:
  - url: https://localhost:5001/api
...
paths:
...
  /questions:
    post:
      tags: [Question]
      summary: Add a new question
      operationId: addQuestion
      requestBody:
        description: Question data to create
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionCreate"
        required: true
      responses:
        "201":
          description: Question created
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [Question]
      summary: List all questions
      operationId: listQuestions
      parameters:
        - name: pageIndex
          in: query
          schema:
            type: integer
            format: int64
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int64
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  paging:
                    $ref: "#/components/schemas/paging"
                  questions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Question"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /questions/{id}:
    get:
      tags: [Question]
      summary: Get a question by ID
      operationId: getQuestionById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Question details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
        "404":
          description: Question not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Question]
      summary: Update an existing question
      operationId: updateQuestion
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        description: Question data to update
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionUpdate"
        required: true
      responses:
        "200":
          description: Question updated
        "400":
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Question not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Question]
      summary: Delete a question
      operationId: deleteQuestion
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Question deleted
        "404":
          description: Question not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /question-options:
    post:
      tags: [QuestionOption]
      summary: Add a new question option
      operationId: addQuestionOption
      requestBody:
        description: Question option data to create
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionOptionCreate"
        required: true
      responses:
        "201":
          description: Question option created
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [QuestionOption]
      summary: List all question options
      operationId: listQuestionOptions
      parameters:
        - name: pageIndex
          in: query
          schema:
            type: integer
            format: int64
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int64
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  paging:
                    $ref: "#/components/schemas/paging"
                  questionOptions:
                    type: array
                    items:
                      $ref: "#/components/schemas/QuestionOption"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /question-options/{id}:
    get:
      tags: [QuestionOption]
      summary: Get a question option by ID
      operationId: getQuestionOptionById
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Question option details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionOption"
        "404":
          description: Question option not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [QuestionOption]
      summary: Update an existing question option
      operationId: updateQuestionOption
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        description: Question option data to update
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionOptionUpdate"
        required: true
      responses:
        "200":
          description: Question option updated
        "400":
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Question option not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [QuestionOption]
      summary: Delete a question option
      operationId: deleteQuestionOption
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Question option deleted
        "404":
          description: Question option not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    paging:
      type: object
      properties:
        pageIndex:
          type: integer
          format: int64
          example: 1
        pageSize:
          type: integer
          format: int64
          example: 10
        totalPages:
          type: integer
          format: int64
          example: 100
        totalRecords:
          type: integer
          format: int64
          example: 100
...
    Question:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        examId:
          type: integer
        questionText:
          type: string
        questionType:
          type: string
        marks:
          type: integer
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        createdBy:
          type: integer
          nullable: true
        updatedBy:
          type: integer
          nullable: true
        topic:
          type: string
          nullable: true
        difficultyLevel:
          type: string
          nullable: true
        isDeleted:
          type: boolean
          nullable: true
    QuestionCreate:
      type: object
      required: [examId, questionText, questionType, marks]
      properties:
        examId:
          type: integer
        questionText:
          type: string
        questionType:
          type: string
        marks:
          type: integer
        topic:
          type: string
          nullable: true
        difficultyLevel:
          type: string
          nullable: true
        createdBy:
          type: integer
          nullable: true
    QuestionUpdate:
      type: object
      required: [examId, questionText, questionType, marks]
      properties:
        examId:
          type: integer
        questionText:
          type: string
        questionType:
          type: string
        marks:
          type: integer
        topic:
          type: string
          nullable: true
        difficultyLevel:
          type: string
          nullable: true
        updatedBy:
          type: integer
          nullable: true
    QuestionOption:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        questionId:
          type: integer
        optionText:
          type: string
        isCorrect:
          type: boolean
        isDeleted:
          type: boolean
          nullable: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        createdBy:
          type: integer
          nullable: true
        updatedBy:
          type: integer
          nullable: true
    QuestionOptionCreate:
      type: object
      required: [questionId, optionText, isCorrect]
      properties:
        questionId:
          type: integer
        optionText:
          type: string
        isCorrect:
          type: boolean
        createdBy:
          type: integer
          nullable: true
    QuestionOptionUpdate:
      type: object
      required: [questionId, optionText, isCorrect]
      properties:
        questionId:
          type: integer
        optionText:
          type: string
        isCorrect:
          type: boolean
        updatedBy:
          type: integer
          nullable: true
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string